fix(windows): remove unnecessary cast in print.rs

+++ src-tauri/src/pdf/print.rs
@@ -120,7 +120,7 @@ fn open_pdf_for_manual_print(path: &Path) -> Result<(), String> {
         fMask: SEE_MASK_NOCLOSEPROCESS,
         lpVerb: windows::core::PCWSTR(operation.as_ptr()),
         lpFile: windows::core::PCWSTR(wide.as_ptr()),
-        nShow: SW_SHOWNORMAL.0 as i32,
+        nShow: SW_SHOWNORMAL.0,
         ..Default::default()
     };
